Rename file burger-menu-nav.vue to BurgerMenuNav.vue
